+++ apiqa_storage/serializers.py
@@ -1,95 +1,86 @@
 import logging
-from rest_framework import serializers
-from rest_framework.exceptions import ValidationError
+from typing import Union
 
-from . import settings
+from django.conf import settings
+from django.utils.translation import ugettext_lazy as _
+from rest_framework import serializers
+
+from .files import FileInfo, file_info
 from .minio_storage import storage
-from .files import file_info
+from .models import Attachment
+from .validators import file_size_validator
 
 logger = logging.getLogger('apiqa-storage')  # noqa
 
 
 __all__ = [
-    'CreateAttachFilesSerializers',
-    'AttachFilesSerializers',
+    'AttachmentSerializer',
+    'AttachmentsSerializerMixin'
 ]
 
 
-class AttachmentField(serializers.FileField):
-    def to_representation(self, data):
-        return {
-            key: value for key, value in data.items()
-            if key in ('uid', 'name', 'size', 'content_type', 'created')
-        }
+def delete_file(attach_file_info: Union[FileInfo, Attachment]):
+    # noinspection PyBroadException
+    try:
+        storage.file_delete(attach_file_info.path)
+    except Exception:  # noqa
+        logger.exception("Delete file failed: %s from bucket: %s",
+                         attach_file_info.path, storage.bucket_name)
 
 
-class AttachFilesSerializers(serializers.Serializer):  # noqa: pylint=abstract-method
-    attachments = serializers.ListField(
-        child=AttachmentField(),
-        max_length=settings.MINIO_STORAGE_MAX_FILES_COUNT,
-        default=list,
-    )
+class AttachmentSerializer(serializers.ModelSerializer):
+    file = serializers.FileField(write_only=True, required=True,
+                                 validators=[file_size_validator])
 
-    def validate_attachments(self, value):  # noqa
-        """
-        Длина имени файла тут не валидируется. Смотри files.py # slugify_name
-        """
-        # Validate files size
-        for attach_file in value:
-            if attach_file.size > settings.MAX_FILE_SIZE:
-                raise ValidationError(
-                    f'Max size of attach file: '
-                    f'{settings.MINIO_STORAGE_MAX_FILE_SIZE}'
-                )
+    class Meta:
+        model = Attachment
+        fields = (
+            'file', 'uid', 'created', 'name', 'size', 'content_type'
+        )
+        read_only_fields = (
+            'uid', 'created', 'name', 'size', 'content_type'
+        )
 
-        return value
-
-
-def upload_files(validated_data: dict):
-    attachments = validated_data.pop('attachments', [])
-
-    attach_files_info = [
-        file_info(attach_file) for attach_file in attachments
-    ]
-
-    # Upload files
-    for attach_file in attach_files_info:
-        # TODO: В середине процесса может случиться ошибка
-        # из хранилки не удалятся загруженные данные
-        storage.file_put(attach_file)
-
-    validated_data['attachments'] = [
-        {
-            'uid': attach_file.uid,
-            'bucket_name': storage.bucket_name,
-            'name': attach_file.name,
-            'created': attach_file.created,
-            'path': attach_file.path,
-            'size': attach_file.size,
-            'content_type': attach_file.content_type,
-        }
-        for attach_file in attach_files_info
-    ]
-    return attach_files_info
-
-
-def delete_files(attach_files_info: list):
-    for attach_file in attach_files_info:
-        # noinspection PyBroadException
-        try:
-            storage.file_delete(attach_file.path)
-        except Exception:  # noqa
-            logger.exception("Delete file failed: %s from bucket: %s",
-                             attach_file.path, storage.bucket_name)
-
-
-class CreateAttachFilesSerializers(AttachFilesSerializers, serializers.ModelSerializer):  # noqa: pylint=abstract-method
     def create(self, validated_data):
-        attach_files_info = upload_files(validated_data)
-
+        user = self.context['request'].user
+        attach_file = validated_data.pop('file')
+        attach_file_info = file_info(attach_file)
+        storage.file_put(attach_file_info)
+        data = {
+            'uid': attach_file_info.uid,
+            'bucket_name': storage.bucket_name,
+            'name': attach_file_info.name,
+            'path': attach_file_info.path,
+            'size': attach_file_info.size,
+            'content_type': attach_file_info.content_type,
+            'user': user
+        }
+        validated_data.update(data)
         try:
             return super().create(validated_data)
         except Exception:
             # Delete files if save model failed
-            delete_files(attach_files_info)
+            delete_file(attach_file_info)
             raise
+
+
+class AttachmentsSerializerMixin(serializers.Serializer):
+    attachments = AttachmentSerializer(many=True, read_only=True)
+    attachment_ids = serializers.PrimaryKeyRelatedField(
+        many=True, write_only=True, queryset=Attachment.objects.all(),
+        source='attachments', required=False
+    )
+
+    def create(self, validated_data):
+        attachments = validated_data.pop('attachments', [])
+        instance = super().create(validated_data)
+        instance.attachments.set(attachments)
+        return instance
+
+    @staticmethod
+    def validate_attachment_ids(value):
+        if len(value) > settings.MINIO_STORAGE_MAX_FILES_COUNT:
+            raise serializers.ValidationError(
+                _('Max files count: %s' % settings
+                  .MINIO_STORAGE_MAX_FILES_COUNT))
+        return value