+++ apiqa_storage/urls.py
@@ -1,11 +1,9 @@
 from django.urls import path
-from apiqa_storage.view import attachment_view
 
+from .routers import router
+from .views import AttachmentView
 
-urlpatterns = [  # noqa
-    path('<uuid:file_uid>',
-        attachment_view,
-        name='attachments',
-        kwargs={'from_user': True},
-    ),
-]
+urlpatterns = [  # noqa: pylint=invalid-name
+    path('<uuid:attachment_uid>', AttachmentView.as_view(),
+         name='user-attachments', kwargs={'from_user': True})
+] + router.urls